+++ itk_dev_shared_components/misc/cpr_util.py
@@ -3,15 +3,15 @@
 from datetime import date
 
 
-def get_birth_date(cpr: str) -> date:
-    """Get the birth date of a person based on their cpr number
+def get_birthdate(cpr: str) -> date:
+    """Get the birthdate of a person based on their cpr number
     using the 7th digit to infer the century.
 
     Args:
         cpr: The cpr number in the format 'ddmmyyxxxx'.
 
     Returns:
-        The birth date based on cpr number.
+        The birthdate based on cpr number.
 
     Raises:
         ValueError: If the given cpr is not in 'ddmmyyxxxx' format.
@@ -67,7 +67,7 @@ def get_age(cpr: str, current_date: date = date.today()) -> int:
     Raises:
         ValueError: If the given cpr is not in 'ddmmyyxxxx' format.
     """
-    birthdate = get_birth_date(cpr)
+    birthdate = get_birthdate(cpr)
 
     age = current_date.year - birthdate.year - ((current_date.month, current_date.day) < (birthdate.month, birthdate.day))